Renaming parsers and lexer
