+++ Cargo.toml
@@ -1,4 +1,3 @@
-workspace = { members = ["lexer"] }
 [package]
 name = "reti"
 version = "0.1.0"
@@ -7,3 +6,6 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+parser={path="./parser"}
+
+workspace = { members = ["parser"] }
