+++ parser/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "lexer"
+name = "parser"
 version = "0.1.0"
 edition = "2021"
